Remove OsuStrainSkill
